refactor: 메뉴를 클래스로 분리
