+++ src/pause.py
@@ -2,6 +2,7 @@ import pygame, sys
 from constant import *
 from setting import Settings
 from button import *
+from menu import Menu
 
 settings = None
 screen = None
@@ -16,48 +17,22 @@ EVENT_RESUME = pygame.event.custom_type()
 '''
 일시 정지 중에 필요한 클래스 정의입니다.
 '''
-class Quit_Menu():
+class Quit_Menu(Menu):
     
     # 가능한 메뉴 목록
     avail_menu = ['돌아가기', '시작 화면', '게임 종료']
 
     # 버튼이 있어야 할 위치 반환
-    get_position = lambda self, index: (self.pos[0]+self.size[0]*1.2*index, self.pos[1])
+    pos_formula = lambda self, i: (self.size[0] * (5 + 10*i) / 30, self.size[1] / 2)
 
-    # 폰트 설정
-    get_font = lambda self, size: pygame.font.Font(RESOURCE_PATH / "font.ttf", size)
+    # x축 정렬 메뉴? y축 정렬 메뉴?
+    axis = "x"
 
     def __init__(self, pos=(0, 0), size=(150, 50)):
-        self.menu = self.avail_menu
-        self.max_menu = len(self.menu)
-        self.button = []
-        self.rect = []
-        self.pos = pos
-        self.size = size
-        self.pressed = False
-
-        # 현재 highlight된 위치의 index
-        self.highlight = 0
-        # 현재 선택된 대상, -1일 경우 마우스 조작 중
-        self.selected = -1
-        self.init_draw()
+        super().__init__(pos, size)
 
     def init_draw(self):
-        self.button = []
-        self.rect = []
-
-        for i in range(self.max_menu):
-            # 버튼 삽입
-            self.button.append(Button(pygame.image.load(RESOURCE_PATH / "main_button.png"),
-                                pos=(self.size[0] * (5 + 10*i) / 30, self.size[1] / 2),
-                                text_input=self.menu[i],
-                                font=self.get_font(50),
-                                base_color="#3a4aab",
-                                hovering_color="White"))
-            # 각 버튼 이벤트 처리용 Rect 삽입
-            self.rect.append(self.button[i].rect)
-            # highlight용 오브젝트 생성
-            self.highlight_obj = pygame.transform.scale(pygame.image.load(RESOURCE_PATH / 'highlight.png'), self.size)
+        super().init_draw()
 
         self.PAUSED_TEXT = self.get_font(45).render("동작을 선택해 주세요.", True, "Black")
         self.PAUSED_RECT = self.PAUSED_TEXT.get_rect(center=(self.size[0] / 2, 260))
@@ -69,14 +44,8 @@ class Quit_Menu():
 
     # 스크린에 자신을 그리기
     def draw(self, screen):
+        super().draw(screen)
         screen.blit(self.PAUSED_TEXT, self.PAUSED_RECT)
-
-        for i in range(self.max_menu):
-            self.button[i].update(screen)
-            if i == self.highlight:
-                self.button[i].forceChangeColor(True, screen)
-            else:
-                self.button[i].forceChangeColor(False, screen)
     
     # 메뉴 선택 시 처리
     def select_menu(self, index):
@@ -89,99 +58,30 @@ class Quit_Menu():
         elif self.avail_menu[index] == '시작 화면':
             pygame.event.post(pygame.event.Event(EVENT_START_MENU)) # 시작 화면으로 복귀
 
-    # 이벤트 처리
-    def handle_event(self, event: pygame.event.Event):
-        for i in range(self.max_menu):
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if self.rect[i].collidepoint(event.pos):
-                    self.select_menu(i)
-                    break # 한 번에 여러 개의 메뉴가 눌리지 않도록 처리
-            elif event.type == pygame.MOUSEMOTION:
-                if self.rect[i].collidepoint(event.pos):
-                    # highlight 대상을 변경
-                    self.highlight = i
-                    # 키보드 선택 해제
-                    self.selected = -1
-            elif event.type == pygame.KEYDOWN:
-                if self.pressed == False:
-                    self.pressed = True
-                    # 엔터 키가 눌렸을 때
-                    if event.key == Settings().settings['enter']:
-                        # 키보드로 선택한 것이 있다면 그 메뉴를 선택
-                        if self.selected != -1: self.select_menu(self.selected)
-                    elif event.key == Settings().settings['left']:
-                        # 선택을 하나 위로 이동
-                        self.selected = self.selected-1 if 0 < self.selected else 0
-                        self.highlight = self.selected
-                    elif event.key == Settings().settings['right']:
-                        # 선택을 하나 아래로 이동
-                        self.selected = self.selected+1 if self.selected < self.max_menu-1 else self.max_menu-1
-                        self.highlight = self.selected
-            # 버튼이 누르고 있어도 계속 동작하지 않게 뗄 때까지는 작동 방지
-            elif event.type == pygame.KEYUP:
-                self.pressed = False
-
-class Paused_Menu():
+class Paused_Menu(Menu):
     
     # 가능한 메뉴 목록
     avail_menu = ['OPTIONS', 'RESUME', 'EXIT']
 
     # 버튼이 있어야 할 위치 반환
-    get_position = lambda self, index: (self.pos[0]+self.size[0]*1.2*index, self.pos[1])
+    pos_formula = lambda self, i: (self.size[0] * (5 + 10*i) / 30, self.size[1] / 2)
 
-    # 폰트 설정
-    get_font = lambda self, size: pygame.font.Font(RESOURCE_PATH / "font.ttf", size)
+    # x축 정렬 메뉴? y축 정렬 메뉴?
+    axis = "x"
 
     def __init__(self, pos=(0, 0), size=(150, 50)):
-        self.menu = self.avail_menu
-        self.max_menu = len(self.menu)
-        self.button = []
-        self.rect = []
-        self.pos = pos
-        self.size = size
-        self.pressed = False
-
-        # 현재 highlight된 위치의 index
-        self.highlight = 0
-        # 현재 선택된 대상, -1일 경우 마우스 조작 중
-        self.selected = -1
-        self.init_draw()
+        super().__init__(pos, size)
 
     def init_draw(self):
-        self.button = []
-        self.rect = []
-
-        for i in range(self.max_menu):
-            # 버튼 삽입
-            self.button.append(Button(pygame.image.load(RESOURCE_PATH / "main_button.png"),
-                                pos=(self.size[0] * (5 + 10*i) / 30, self.size[1] / 2),
-                                text_input=self.menu[i],
-                                font=self.get_font(50),
-                                base_color="#3a4aab",
-                                hovering_color="White"))
-            # 각 버튼 이벤트 처리용 Rect 삽입
-            self.rect.append(self.button[i].rect)
-            # highlight용 오브젝트 생성
-            self.highlight_obj = pygame.transform.scale(pygame.image.load(RESOURCE_PATH / 'highlight.png'), self.size)
+        super().init_draw()
 
         self.PAUSED_TEXT = self.get_font(45).render("Game is paused.", True, "Black")
         self.PAUSED_RECT = self.PAUSED_TEXT.get_rect(center=(self.size[0] / 2, 260))
 
-    # 크기 변경에 맞춰 재조정
-    def resize(self, size):
-        self.size = size
-        self.init_draw()
-
     # 스크린에 자신을 그리기
     def draw(self, screen):
+        super().draw(screen)
         screen.blit(self.PAUSED_TEXT, self.PAUSED_RECT)
-
-        for i in range(self.max_menu):
-            self.button[i].update(screen)
-            if i == self.highlight:
-                self.button[i].forceChangeColor(True, screen)
-            else:
-                self.button[i].forceChangeColor(False, screen)
     
     # 메뉴 선택 시 처리
     def select_menu(self, index):
@@ -193,38 +93,6 @@ class Paused_Menu():
             pygame.event.post(pygame.event.Event(EVENT_OPEN_OPTION)) # 옵션 열기
         elif self.avail_menu[index] == 'RESUME':
             pygame.event.post(pygame.event.Event(EVENT_RESUME)) # 일시정지 해제
-
-    # 이벤트 처리
-    def handle_event(self, event: pygame.event.Event):
-        for i in range(self.max_menu):
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if self.rect[i].collidepoint(event.pos):
-                    self.select_menu(i)
-                    break # 한 번에 여러 개의 메뉴가 눌리지 않도록 처리
-            elif event.type == pygame.MOUSEMOTION:
-                if self.rect[i].collidepoint(event.pos):
-                    # highlight 대상을 변경
-                    self.highlight = i
-                    # 키보드 선택 해제
-                    self.selected = -1
-            elif event.type == pygame.KEYDOWN:
-                if self.pressed == False:
-                    self.pressed = True
-                    # 엔터 키가 눌렸을 때
-                    if event.key == Settings().settings['enter']:
-                        # 키보드로 선택한 것이 있다면 그 메뉴를 선택
-                        if self.selected != -1: self.select_menu(self.selected)
-                    elif event.key == Settings().settings['left']:
-                        # 선택을 하나 위로 이동
-                        self.selected = self.selected-1 if 0 < self.selected else 0
-                        self.highlight = self.selected
-                    elif event.key == Settings().settings['right']:
-                        # 선택을 하나 아래로 이동
-                        self.selected = self.selected+1 if self.selected < self.max_menu-1 else self.max_menu-1
-                        self.highlight = self.selected
-            # 버튼이 누르고 있어도 계속 동작하지 않게 뗄 때까지는 작동 방지
-            elif event.type == pygame.KEYUP:
-                self.pressed = False
 
 '''
 일시 정지 중에 메인 루프를 대체하여 작동하는 루프입니다.
